moved config util code under util
