+++ packages/node/src/remote/client.ts
@@ -3,14 +3,11 @@ import {
   FetchError,
   SdkEvaluationApi,
 } from '@amplitude/experiment-core';
+import { populateRemoteConfigDefaults } from 'src/util/config';
 
 import { version as PACKAGE_VERSION } from '../../gen/version';
 import { FetchHttpClient, WrapperClient } from '../transport/http';
-import {
-  ExperimentConfig,
-  RemoteEvaluationConfig,
-  populateRemoteConfigDefaults,
-} from '../types/config';
+import { ExperimentConfig, RemoteEvaluationConfig } from '../types/config';
 import { FetchOptions } from '../types/fetch';
 import { ExperimentUser } from '../types/user';
 import { Variant, Variants } from '../types/variant';
